web/requirements.txt
Django==2.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
gunicorn==19.3.0
psycopg2==2.6
pytz==2018.5
redis==2.10.3
